[WIP] Add a slideshow in the login page.
The slideshow is implemented with the Django Hero Slider app:
https://github.com/bitlabstudio/django-hero-slider
We set the version of Django to 1.10.* to avoid using the last version
(1.11) because Django Hero Slider needs easy_thumbnails, which is not
compatible yet with Django 1.11.

Add the Django admin site and modify the authentication backends to
allow login to it (the items of the slideshow will be created
dynamically in the Django admin site).

Add the media/ directory to store the images of the slideshow.

(cherry picked from commit 3e96dd0cb2d99da271386664521118a264c6406c)